Replace shell command string with Python-based action (for compatibility when building on Windows)
