+++ qcrbox/qcrbox/cli/helpers/__init__.py
@@ -6,8 +6,10 @@ from .cli_helpers import add_cli_option_enable_disable_components, add_verbose_o
 from .docker_project import DockerProject
 from .doit_helpers import make_task, run_tasks
 from .qcrbox_helpers import (
+    QCrBoxSubprocessError,
     get_mkdocs_config_file_path,
     get_repo_root,
+    prettyprint_called_process_error,
 )
