+++ qcrbox/qcrbox/cli/helpers/qcrbox_helpers.py
@@ -3,6 +3,7 @@
 import functools
 import shutil
 import subprocess
+import textwrap
 from pathlib import Path
 from typing import Optional, TypeVar
 
@@ -67,3 +68,29 @@ def find_common_repo_root(*files: PathLike):
 
 def get_mkdocs_config_file_path():
     return get_repo_root().joinpath("mkdocs.yml")
+
+
+class QCrBoxSubprocessError(Exception):
+    """
+    Custom exception to indicate errors during the build process of QCrBox components.
+    """
+
+
+def prettyprint_called_process_error(exc: subprocess.CalledProcessError):
+    cmd = " ".join(exc.cmd)
+    prefix = " " * 24
+    captured_stdout = textwrap.indent(f"\n\n{exc.stdout.decode()}\n" if exc.stdout else "(not captured)", prefix=prefix)
+    captured_stderr = textwrap.indent(f"\n\n{exc.stderr.decode()}\n" if exc.stderr else "(not captured)", prefix=prefix)
+    msg = textwrap.dedent(
+        f"""\
+        An error occurred when executing the following command:
+
+            {cmd}
+
+        Return code: {exc.returncode}
+
+        Captured stdout: {captured_stdout}
+        Captured stderr: {captured_stderr}
+        """
+    )
+    return msg
