+++ qcrbox/qcrbox/cli/subcommands/build.py
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: MPL-2.0
 import shutil
+import subprocess
 from pathlib import Path
 from typing import Iterable
 
@@ -9,10 +10,12 @@ from loguru import logger
 
 from ..helpers import (
     DockerProject,
+    QCrBoxSubprocessError,
     add_cli_option_enable_disable_components,
     add_verbose_option,
     get_repo_root,
     make_task,
+    prettyprint_called_process_error,
     run_tasks,
 )
 
@@ -56,6 +59,25 @@ def make_action_to_copy_file(src, dest):
     return action_copy_file
 
 
+def make_action_to_build_qcrbox_wheel(base_ancestor_qcrbox_dist_dir):
+    repo_root = get_repo_root()
+    qcrbox_package_root = repo_root.joinpath("qcrbox")
+
+    try:
+        cmd = [shutil.which("hatch"), "build", "-t", "wheel", base_ancestor_qcrbox_dist_dir]
+        proc = subprocess.run(cmd, cwd=qcrbox_package_root, shell=False, check=False, capture_output=False)
+    except Exception as exc:
+        raise QCrBoxSubprocessError(f"Error when trying to run docker compose command: {exc}")
+
+    try:
+        proc.check_returncode()
+    except subprocess.CalledProcessError as exc:
+        error_msg = prettyprint_called_process_error(exc)
+        raise QCrBoxSubprocessError(error_msg)
+
+    return proc
+
+
 @make_task
 def task_clone_qcrboxtools_repo(dry_run: bool):
     qcrboxtools_repo_url = "https://github.com/QCrBox/QCrBoxTools.git"
@@ -71,7 +93,6 @@ def task_clone_qcrboxtools_repo(dry_run: bool):
 @make_task
 def task_build_qcrbox_python_package(dry_run: bool):
     repo_root = get_repo_root()
-    qcrbox_package_root = repo_root.joinpath("qcrbox")
 
     action_descr = "Building" if not dry_run else "Would build"
     actions = [lambda: logger.info(f"{action_descr} Python package: qcrbox", dry_run=dry_run)]
@@ -79,7 +100,7 @@ def task_build_qcrbox_python_package(dry_run: bool):
         base_ancestor_qcrbox_dist_dir = repo_root.joinpath("services/base_images/base_ancestor/qcrbox_dist/")
         requirements_files = list(repo_root.glob("qcrbox/requirements*.txt"))
 
-        action_build_qcrbox_wheel = f"cd {qcrbox_package_root} && hatch build -t wheel {base_ancestor_qcrbox_dist_dir}"
+        action_build_qcrbox_wheel = make_action_to_build_qcrbox_wheel(base_ancestor_qcrbox_dist_dir)
         actions_copy_requirements_files = [
             make_action_to_copy_file(filename, base_ancestor_qcrbox_dist_dir) for filename in requirements_files
         ]
